Move "SearchSettings" into "LibrarySettings"
